chore: update sequentialnumbercounter id column to bigInt

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/reservedvalue/SequentialNumberCounter.java
@@ -32,7 +32,7 @@ package org.hisp.dhis.reservedvalue;
  */
 public class SequentialNumberCounter
 {
-    private int id;
+    private long id;
 
     private String ownerUid;
 
@@ -52,12 +52,12 @@ public class SequentialNumberCounter
         this.counter = counter;
     }
 
-    public int getId()
+    public long getId()
     {
         return id;
     }
 
-    public void setId( int id )
+    public void setId( long id )
     {
         this.id = id;
     }
